Refactor context managers into separate module
